requirements.txt
ffmpeg-python==0.2.0
groq==0.14.0
langchain==0.3.14
langchain-groq==0.2.3
langgraph==0.2.61
pywhispercpp==1.3.0
